更新 Install 脚本

+++ install/docker.sh
@@ -1,19 +1,22 @@
 #!/bin/bash
 [ -f "/usr/bin/docker" ] && echo 'Docker Installed' && exit
 
-# 更换软件源
-sudo cp /etc/apt/sources.list /etc/apt/sources.list.bak`date +%Y%m%d%H%M` \
-  && sudo sed -i 's/archive.ubuntu.com/mirrors.ustc.edu.cn/g' /etc/apt/sources.list
-
+# https://docs.docker.com/engine/install/ubuntu/
 # Update the apt
 sudo apt-get update && sudo apt-get install -y ca-certificates curl gnupg lsb-release
 
 # Add Docker’s official GPG key:
-curl -fsSL https://mirrors.ustc.edu.cn/docker-ce/linux/ubuntu/gpg | sudo gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
+sudo mkdir -p /etc/apt/keyrings
+curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+if [ ! -f "/etc/apt/keyrings/docker.gpg" ]; then
+    echo 'Error: Add Docker GPG key failed.'
+    exit
+fi
+sudo chmod a+r /etc/apt/keyrings/docker.gpg
 
 # Use the following command to set up the stable repository
 echo \
-  "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://mirrors.ustc.edu.cn/docker-ce/linux/ubuntu \
+  "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
   $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
 
 # Install Docker Engine
@@ -27,5 +30,9 @@ echo \
 }" | sudo tee /etc/docker/daemon.json > /dev/null
 
 # Start docker
-sudo service docker start
-echo 'Install Docker Done.'
+if [ -f "/usr/bin/docker" ]; then
+    sudo service docker start
+    echo 'Install Docker Done.'
+else
+    echo 'Error: Install Docker failed.'
+fi
